Fix xvfb bin; use init path var
This commit fixes the default path to the xvfb binary on RedHat and
Debian, which is capitalized (Xvfb vs xvfb).

This also makes use of the display::params::init_path variable to
properly place the init script on debian for xvfb.

Also use the $service variable for the PID filename and service name.