Parameterize display for display::env
This commit parameterizes the 'display' variable for the 'env' class,
which can be used for adding an environment variable for the X display.

Prior to this, the display was obtained from the base class, which did
not allow this class to be used independently.  This defaults the value
to '0', which is obtained from 'params'.

When declaring the base class (display), whatever value the 'display'
parameter has will be used by 'env'.