Add support for custom xvfb arguments
This commit introduces the ability to pass custom arguments for starting
xvfb.  It adds a "custom_args" parameter to the display::xvfb class and
"xvfb_custom_args" to the base class.

Templates for the init scripts are modified to pass the custom
arguments, if set.

When custom arguments are passed, all of the other parameters are
ignored.